+++ src/Management.Interest/Management.Interest.Infrastructure.Data.Query/Queries/v1/GetInterestRate/GetInterestRateQueryResponse.cs
@@ -5,7 +5,7 @@ namespace Management.Interest.Infrastructure.Data.Query.Queries.v1.GetInterestRa
     public class GetInterestRateQueryResponse
     {
         [JsonProperty("interestRate")]
-        public decimal InterestRate { get; set; }
+        public double InterestRate { get; set; }
 
         [JsonProperty("formattedInterestRate")]
         public string FormattedInterestRate { get; set; }
